Move the mixin to the mixins directory and move tests as well. Move the association table to models in resource_file_collection_association.py Renamed for "...assoc" to "...association" for consistency with other association tables.
