+++ tethysext/atcore/mixins/__init__.py
@@ -3,3 +3,4 @@ from tethysext.atcore.mixins.attributes_mixin import AttributesMixin  # noqa: F4
 from tethysext.atcore.mixins.options_mixin import OptionsMixin  # noqa: F401
 from tethysext.atcore.mixins.results_mixin import ResultsMixin  # noqa: F401
 from tethysext.atcore.mixins.user_lock_mixin import UserLockMixin  # noqa: F401
+from tethysext.atcore.mixins.file_collection_mixin import FileCollectionMixin  # noqa: F401
