+++ tethysext/atcore/models/file_database/__init__.py
@@ -8,5 +8,4 @@
 """
 from tethysext.atcore.models.file_database.file_collection import FileCollection  # noqa: F401, E402, E501
 from tethysext.atcore.models.file_database.file_database import FileDatabase  # noqa: F401, E402, E501
-from tethysext.atcore.models.file_database.file_collection_mixin import resource_file_collection_assoc, \
-    FileCollectionMixin  # noqa: F401, E402, E501
+from tethysext.atcore.models.file_database.resource_file_collection_association import resource_file_collection_association  # noqa: F401, E402, E501
